Solved circular imports with forward references & module renaming
