+++ src/tsdf/read_binary.py
@@ -8,19 +8,15 @@ import os
 from typing import Any, Dict
 import numpy as np
 from tsdf.numpy_utils import (
-    data_type_numpy_to_tsdf,
     data_type_tsdf_to_numpy,
-    bits_numpy_to_tsdf,
     bytes_tsdf_to_numpy,
-    endianness_numpy_to_tsdf,
     endianness_tsdf_to_numpy,
-    rows_numpy_to_tsdf,
 )
-from tsdf.tsdf_metadata import TSDFMetadata
+from tsdf import tsdfmetadata
 
 
 def load_binary_from_metadata(
-    metadata_dir: str, metadata: TSDFMetadata, start_row: int = 0, end_row: int = -1
+    metadata_dir: str, metadata: 'tsdfmetadata.TSDFMetadata', start_row: int = 0, end_row: int = -1
 ) -> np.ndarray:
     """
     Use metadata properties to load and return numpy array from a binary file.
@@ -32,7 +28,7 @@ def load_binary_from_metadata(
 
     :return: numpy array containing the data."""
     bin_path = os.path.join(metadata_dir, metadata.file_name)
-    return load_binary_file(
+    return _load_binary_file(
         bin_path,
         metadata.data_type,
         metadata.bits,
@@ -44,7 +40,7 @@ def load_binary_from_metadata(
     )
 
 
-def load_binary_file(
+def _load_binary_file(
     bin_file_path: str,
     data_type: str,
     n_bits: int,
@@ -89,42 +85,3 @@ def load_binary_file(
         raise Exception("Number of rows doesn't match file length.")
 
     return values
-
-
-def get_metadata_from_ndarray(data: np.ndarray) -> Dict[str, Any]:
-    """
-    Retrieve metadata information encoded in the NumPy array.
-
-    :param data: NumPy array containing the data.
-
-    :return: dictionary containing the metadata.
-    """
-
-    metadata = {
-        "data_type": data_type_numpy_to_tsdf(data),
-        "bits": bits_numpy_to_tsdf(data),
-        "endianness": endianness_numpy_to_tsdf(data),
-        "rows": rows_numpy_to_tsdf(data),
-    }
-    return metadata
-
-
-def write_binary_file(
-    file_dir: str, file_name: str, data: np.ndarray, metadata: dict
-) -> TSDFMetadata:
-    """
-    Save binary file based on the provided NumPy array.
-
-    :param file_dir: path to the directory where the file will be saved.
-    :param file_name: name of the file to be saved.
-    :param data: NumPy array containing the data.
-    :param metadata: dictionary containing the metadata.
-
-    :return: TSDFMetadata object.
-    """
-    path = os.path.join(file_dir, file_name)
-    data.tofile(path)
-    metadata.update(get_metadata_from_ndarray(data))
-    metadata.update({"file_name": file_name})
-
-    return TSDFMetadata(metadata, file_dir)
