+++ src/tsdf/tsdfmetadata.py
@@ -1,7 +1,7 @@
 import copy
 from typing import Any, Dict, List
 from numpy import ndarray
-import tsdf.io_metadata as io_metadata
+from tsdf import parse_metadata, read_binary
 
 class TSDFMetadataFieldError(Exception):
     "Raised when the TSDFMetadata is missing an obligatory field."
@@ -60,7 +60,7 @@ class TSDFMetadata:
         :param dir_path: path to the directory where the metadata file is stored.
         :param file_name: (optional) name of the metadata file.
         """
-        io_metadata.contains_tsdf_mandatory_fields(dictionary)
+        parse_metadata.contains_tsdf_mandatory_fields(dictionary)
         for key, value in dictionary.items():
             setattr(self, key, value)
         self.file_dir_path = dir_path
@@ -85,5 +85,4 @@ class TSDFMetadata:
 
         :return: binary file as a numpy array.
         """
-        import tsdf.io_binary as io_binary
-        return io_binary.load_binary_from_metadata(self.file_dir_path, self)
+        return read_binary.load_binary_from_metadata(self.file_dir_path, self)
