+++ tests/test_legacy_tsdf_utils.py
@@ -1,7 +1,7 @@
 import os
 import unittest
-from tsdf.load_tsdf import load_metadata_from_path
-from tsdf.legacy_tsdf_utils import generate_tsdf_metadata_from_tsdb
+from tsdf import read_tsdf
+from tsdf import legacy_tsdf_utils
 from tsdf.constants import TestConstants as CONST
 
 
@@ -19,13 +19,13 @@ class TestConversion(unittest.TestCase):
         )
 
         # Generate a TSDF metadata file from TSDB
-        generate_tsdf_metadata_from_tsdb(path_to_file, path_to_new_file)
+        legacy_tsdf_utils.generate_tsdf_metadata_from_tsdb(path_to_file, path_to_new_file)
 
         # Load the generated metadata file
-        new_meta = load_metadata_from_path(path_to_new_file)
+        new_meta = read_tsdf.load_metadata_from_path(path_to_new_file)
 
         # Load the existing metadata file
-        existing_meta = load_metadata_from_path(path_to_existing_tsdf_file)
+        existing_meta = read_tsdf.load_metadata_from_path(path_to_existing_tsdf_file)
 
         # Compare the two metadata files (whether the mapped TSDFs fields are the same)
         self.assertEqual(
